third_party/WebKit/LayoutTests/accessibility/deleting-iframe-destroys-axcache-expected.txt

Before

After

End of test

This tests that deleting an iframe doesn't cause the accessibility cache to be destroyed and recreated.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE

Before:
AXRole: AXWebArea
    AXRole: AXParagraph
        AXRole: AXStaticText "Before"
            AXRole: AXInlineTextBox "Before"
    AXRole: AXGroup
        AXRole: AXUnknown
            AXRole: AXScrollArea
                AXRole: AXWebArea
                    AXRole: AXGroup
                        AXRole: AXButton "Click me"
    AXRole: AXParagraph
        AXRole: AXStaticText "After"
            AXRole: AXInlineTextBox "After"
    AXRole: AXParagraph
        AXRole: AXStaticText "End of test"

After:
AXRole: AXWebArea
    AXRole: AXParagraph
        AXRole: AXStaticText "Before"
            AXRole: AXInlineTextBox "Before"
    AXRole: AXParagraph
        AXRole: AXStaticText "After"
            AXRole: AXInlineTextBox "After"
    AXRole: AXParagraph
        AXRole: AXStaticText "End of test"

PASS frameBodyRole == frameBody.role is false
PASS frameGroupRole == frameGroup.role is false
PASS frameButtonRole == frameButton.role is false
PASS before.isEqual(newBefore) is true
PASS after.isEqual(newAfter) is true

TEST COMPLETE
